+++ datar/datar/__init__.py
@@ -1,4 +1,4 @@
 """Specific verbs/funcs from this package"""
 
 from .verbs import get, flatten
-from .funcs import itemgetter
+from .funcs import itemgetter, attrgetter, pd_str, pd_cat, pd_dt